web/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
APScheduler==3.10.4
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
